/verify/assert/
