+++ cli_tools/common/assert/assert.go
@@ -12,8 +12,23 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-// Package verify provides convenience functions that help a method or constructor check whether
-// it was invoked correctly (whether its preconditions have been met).
-//
-// When the verification fails, a panic occurs, allowing bugs to be found in a fail-fast manner.
-package verify
+package assert
+
+import "fmt"
+
+// GreaterThanOrEqualTo asserts that value is greater than or equal to limit.
+func GreaterThanOrEqualTo(value int, limit int) {
+	if value < limit {
+		panic(fmt.Sprintf("Expected %d >= %d", value, limit))
+	}
+}
+
+// Contains asserts that element is a member of arr.
+func Contains(element string, arr []string) {
+	for _, e := range arr {
+		if e == element {
+			return
+		}
+	}
+	panic(fmt.Sprintf("%s is not a member of %v", element, arr))
+}
