+++ cli_tools/common/assert/doc.go
@@ -12,25 +12,8 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package verify
-
-import "fmt"
-
-// GreaterThanOrEqualTo verifies that value is greater than or equal to limit,
-// and panics if the verification fails.
-func GreaterThanOrEqualTo(value int, limit int) {
-	if value < limit {
-		panic(fmt.Sprintf("Expected %d >= %d", value, limit))
-	}
-}
-
-// Contains verifies that element is a member of arr, and panics if the verification fails.
-func Contains(element string, arr []string) {
-	for _, e := range arr {
-		if e == element {
-			return
-
-		}
-	}
-	panic(fmt.Sprintf("%s is not a member of %v", element, arr))
-}
+// Package assert provides convenience functions that help a method or constructor check whether
+// it was invoked correctly (whether its preconditions have been met).
+//
+// When the assertion fails, a panic occurs, allowing bugs to be found in a fail-fast manner.
+package assert
